Created an actual api route for getting a location's weather
